Move all other params to region class
